@@ -4,9 +4,10 @@
   "version": "1.3.3",
   "description": "An app that creates xml files for Razer Synapse Macro.",
   "main": "main.js",
+  "type": "module",
   "devDependencies": {
-    "electron": "^9.4.4",
-    "electron-builder": "^22.11.7"
+    "electron": "^32.0.1",
+    "electron-builder": "^24.13.3"
   },
   "scripts": {
     "start": "electron .",
@@ -38,7 +39,7 @@
   "author": "Leonard Mueller",
   "license": "ISC",
   "dependencies": {
-    "electron-store": "^8.0.0",
-    "electron-updater": "^4.3.9"
+    "electron-store": "^8.2.0",
+    "electron-updater": "^6.2.1"
   }
 }
